chore: `pytest` boilerplate
